rename : Product내부의 model 패키지의 이름을 domain으로 변경
