+++ src/main/java/gift/product/domain/Product.java
@@ -1,4 +1,4 @@
-package gift.product.model;
+package gift.product.domain;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
